Gallerysearch + improvements

+++ category_delete.php
@@ -2,9 +2,9 @@
 require_once 'autoloader.php';
 AuthService::InitAuth();
 
-$categoryRepository = new CategoryRepository();
-if(!empty($_GET['id']) && $categoryRepository->does_category_exist($_GET['id']) && $_SESSION['is_admin'] == 1) {
-    if($categoryRepository->get_article_count_by_category($_GET['id'])['article_count'] == 0) {
+if(!empty($_GET['id']) && $_SESSION['is_admin'] == 1) {
+    $categoryRepository = new CategoryRepository();
+    if($categoryRepository->does_category_exist($_GET['id']) && $categoryRepository->get_article_count_by_category($_GET['id'])['article_count'] == 0) {
         $categoryRepository->delete_category($_GET['id']);
     }
 }
